requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0
accelerate>=0.20.0
evaluate>=0.4.0
peft>=0.4.0  # For LoRA fine-tuning
sentencepiece>=0.1.99
sacremoses>=0.0.53
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1
seqeval>=1.2.2  # For token classification evaluation
pyabsa>=2.0.0  # PyABSA library
conllu>=4.5.2  # For CoNLL-U format
xmltodict>=0.13.0  # For XML dataset parsing
lxml>=4.9.0
pyyaml>=6.0
tensorboard>=2.13.0
jupyter>=1.0.0
ipywidgets>=8.0.0
plotly>=5.15.0
networkx>=3.1  # For aspect relationship visualization
wordcloud>=1.9.0
torchcrf>=0.7.2  # For CRF layer in token classification